+++ ERA/src/core/transaction/BuyNameTransaction.java
@@ -336,7 +336,7 @@ public class BuyNameTransaction extends Transaction {
     }
 
     @Override
-    public int calcBaseFee() {
+    public long calcBaseFee() {
         return calcCommonFee();
     }
 }
