+++ ERA/src/core/transaction/DeployATTransaction.java
@@ -522,7 +522,7 @@ public class DeployATTransaction extends Transaction {
     }
 
     @Override
-    public int calcBaseFee() {
+    public long calcBaseFee() {
         return calcCommonFee();
     }
 }
