+++ ERA/src/core/transaction/IssueStatementRecord.java
@@ -485,7 +485,7 @@ public class IssueStatementRecord extends Transaction {
     }
 
     @Override
-    public int calcBaseFee() {
+    public long calcBaseFee() {
         return calcCommonFee();
     }
 }
