+++ ERA/src/core/transaction/R_SignNote.java
@@ -658,7 +658,7 @@ public class R_SignNote extends Transaction {
     }
 
     @Override
-    public int calcBaseFee() {
+    public long calcBaseFee() {
         return calcCommonFee();
     }
